+++ tests/e2e/core-tests/specs/index.js
@@ -37,6 +37,7 @@ const runProductSearchTest = require( './merchant/wp-admin-product-search.test' 
 const runMerchantOrdersCustomerPaymentPage = require( './merchant/wp-admin-order-customer-payment-page.test' );
 const runMerchantOrderEmailsTest = require( './merchant/wp-admin-order-emails.test' );
 const runOrderSearchingTest = require( './merchant/wp-admin-order-searching.test' );
+const runAnalyticsPageLoadsTest = require( './merchant/wp-admin-analytics-page-loads.test' );
 
 // REST API tests
 const runExternalProductAPITest = require( './api/external-product.test' );
@@ -81,6 +82,7 @@ const runMerchantTests = () => {
 	runProductEditDetailsTest();
 	runProductSearchTest();
 	runMerchantOrdersCustomerPaymentPage();
+	runAnalyticsPageLoadsTest();
 }
 
 const runApiTests = () => {
@@ -129,5 +131,6 @@ module.exports = {
 	runAddNewShippingZoneTest,
 	runProductBrowseSearchSortTest,
 	runApiTests,
+	runAnalyticsPageLoadsTest,
 	runCheckoutCreateAccountTest,
 };