+++ tests/e2e/utils/src/flows/merchant.js
@@ -20,6 +20,7 @@ const {
 	WP_ADMIN_SETUP_WIZARD,
 	WP_ADMIN_WC_SETTINGS,
 	WP_ADMIN_NEW_SHIPPING_ZONE,
+	WP_ADMIN_ANALYTICS_PAGES,
 	WP_ADMIN_ALL_USERS_VIEW,
 } = require( './constants' );
 
@@ -180,6 +181,12 @@ const merchant = {
 
 	openEmailLog: async () => {
 		await page.goto( `${baseUrl}wp-admin/tools.php?page=wpml_plugin_log`, {
+			waitUntil: 'networkidle0',
+		} );
+	},
+
+	openAnalyticsPage: async ( pageName ) => {
+		await page.goto( WP_ADMIN_ANALYTICS_PAGES + pageName, {
 			waitUntil: 'networkidle0',
 		} );
 	},